+++ packages/frontend/tests/performance/k6/helpers/config.js
@@ -1,3 +1,7 @@
+/*
+ * Get the base URL for arbeidsflate (AF) based on the environment.
+ * @returns {string} - The base URL for AF.
+ */
 export const afUrl = ( function () {
     switch (__ENV.ENVIRONMENT) {
       case 'yt': return 'https://af.yt.altinn.cloud/';
