+++ packages/frontend/tests/performance/k6/helpers/token.js
@@ -1,3 +1,7 @@
+/**
+ * This module provides functions to fetch tokens from the Altinn Test Tools Token Generator API.
+ * It supports both enterprise and personal tokens, caching them to avoid unnecessary requests.
+ */
 import http from "k6/http";
 import { URL } from 'https://jslib.k6.io/url/1.0.0/index.js';
 import encoding from "k6/encoding";
