+++ packages/frontend/tests/performance/k6/testData/usersWithDialogs-tt.csv
@@ -77,7 +77,6 @@ pid
 04918296358
 20815897634
 42865800309
-01476901207
 20905995783
 16928299868
 27834496882
@@ -205,7 +204,6 @@ pid
 20860249716
 29900699229
 29906299990
-20042530213
 17902647988
 17862348044
 57887800482
@@ -228,7 +226,6 @@ pid
 04889498663
 04877098058
 04825899710
-03099324324
 01908498565
 01836796310
 01026200969
@@ -251,7 +248,6 @@ pid
 22839773718
 22837598722
 21818797394
-20902194100
 20884599432
 02848597905
 67850575964
